refactor: SessionRepository 디렉토리 경로 수정
