+++ src/main/java/nextstep/courses/service/CourseService.java
@@ -2,7 +2,7 @@ package nextstep.courses.service;
 
 import nextstep.courses.domain.Course;
 import nextstep.courses.domain.CourseRepository;
-import nextstep.courses.domain.SessionRepository;
+import nextstep.courses.domain.session.SessionRepository;
 import nextstep.courses.domain.session.Session;
 import nextstep.courses.domain.session.SessionDescriptor;
 import nextstep.courses.domain.session.Sessions;
